all: converted views.py to seperate class-based views
